Explicit cast

+++ avalon/tools/loader/model.py
@@ -278,8 +278,8 @@ class SubsetsModel(TreeModel):
         repre_info = item.get("repre_info")
         if repre_info:
             repres = "{}/{}".format(
-                math.floor(float(repre_info['avail_repre'])),
-                math.floor(float(repre_info['repre_count'])))
+                int(math.floor(float(repre_info['avail_repre']))),
+                int(math.floor(float(repre_info['repre_count']))))
             item["repre_info"] = repres
             item["repre_icon"] = self.repre_icons.get(repre_info["provider"])
